+++ front/lib/actions/mcp_internal_actions/constants.ts
@@ -30,6 +30,7 @@ import { GOOGLE_SHEETS_SERVER } from "@app/lib/api/actions/servers/google_sheets
 import { HUBSPOT_SERVER } from "@app/lib/api/actions/servers/hubspot/metadata";
 import { IMAGE_GENERATION_SERVER } from "@app/lib/api/actions/servers/image_generation/metadata";
 import { INCLUDE_DATA_SERVER } from "@app/lib/api/actions/servers/include_data/metadata";
+import { MICROSOFT_EXCEL_SERVER } from "@app/lib/api/actions/servers/microsoft_excel/metadata";
 import { MISSING_ACTION_CATCHER_SERVER } from "@app/lib/api/actions/servers/missing_action_catcher/metadata";
 import { MONDAY_SERVER } from "@app/lib/api/actions/servers/monday/metadata";
 import { NOTION_SERVER } from "@app/lib/api/actions/servers/notion/metadata";
@@ -387,7 +388,6 @@ export const INTERNAL_MCP_SERVERS = {
     allowMultipleInstances: false,
     isRestricted: undefined,
     isPreview: false,
-    tools_stakes: undefined,
     tools_arguments_requiring_approval: undefined,
     tools_retry_policies: undefined,
     timeoutMs: undefined,
@@ -399,13 +399,13 @@ export const INTERNAL_MCP_SERVERS = {
     allowMultipleInstances: true,
     isRestricted: undefined,
     isPreview: false,
-    metadata: SLACK_PERSONAL_SERVER,
     tools_arguments_requiring_approval: {
       post_message: ["channel"],
       schedule_message: ["channel"],
     },
     tools_retry_policies: undefined,
     timeoutMs: undefined,
+    metadata: SLACK_PERSONAL_SERVER,
   },
   google_sheets: {
     id: 19,
@@ -439,7 +439,6 @@ export const INTERNAL_MCP_SERVERS = {
     allowMultipleInstances: false,
     isRestricted: undefined,
     isPreview: false,
-    tools_stakes: undefined,
     tools_arguments_requiring_approval: undefined,
     tools_retry_policies: undefined,
     timeoutMs: undefined,
@@ -857,30 +856,20 @@ export const INTERNAL_MCP_SERVERS = {
     allowMultipleInstances: true,
     isRestricted: undefined,
     isPreview: false,
-    tools_stakes: {
-      list_excel_files: "never_ask",
-      get_worksheets: "never_ask",
-      read_worksheet: "never_ask",
-      write_worksheet: "high",
-      create_worksheet: "low",
-      clear_range: "high",
-    },
     tools_arguments_requiring_approval: undefined,
     tools_retry_policies: undefined,
     timeoutMs: undefined,
-    serverInfo: {
-      name: "microsoft_excel",
-      version: "1.0.0",
-      description: "Work with Excel files in SharePoint.",
-      authorization: {
-        provider: "microsoft_tools" as const,
-        supported_use_cases: ["personal_actions"] as const,
-        scope:
-          "User.Read Files.ReadWrite.All Sites.Read.All offline_access" as const,
+    metadata: {
+      ...MICROSOFT_EXCEL_SERVER,
+      serverInfo: {
+        ...MICROSOFT_EXCEL_SERVER.serverInfo,
+        authorization: {
+          provider: "microsoft_tools" as const,
+          supported_use_cases: ["personal_actions"] as const,
+          scope:
+            "User.Read Files.ReadWrite.All Sites.Read.All offline_access" as const,
+        },
       },
-      icon: "MicrosoftExcelLogo",
-      documentationUrl: null,
-      instructions: null,
     },
   },
   http_client: {
@@ -1216,7 +1205,6 @@ export const INTERNAL_MCP_SERVERS = {
     allowMultipleInstances: false,
     isRestricted: undefined,
     isPreview: false,
-    tools_stakes: undefined,
     tools_arguments_requiring_approval: undefined,
     tools_retry_policies: undefined,
     timeoutMs: undefined,
@@ -1230,7 +1218,6 @@ export const INTERNAL_MCP_SERVERS = {
     allowMultipleInstances: false,
     isRestricted: undefined,
     isPreview: false,
-    tools_stakes: undefined,
     tools_arguments_requiring_approval: undefined,
     tools_retry_policies: undefined,
     timeoutMs: undefined,
@@ -1464,13 +1451,6 @@ export const INTERNAL_MCP_SERVERS = {
     isRestricted: ({ featureFlags }) => {
       return !featureFlags.includes("projects");
     },
-    tools_stakes: {
-      list_project_files: "never_ask",
-      add_project_file: "high",
-      update_project_file: "high",
-      delete_project_file: "high",
-      read_project_journal_entry: "never_ask",
-    },
     tools_arguments_requiring_approval: undefined,
     tools_retry_policies: undefined,
     timeoutMs: undefined,
@@ -1484,10 +1464,10 @@ export const INTERNAL_MCP_SERVERS = {
     isRestricted: ({ featureFlags }) => {
       return !featureFlags.includes("agent_builder_copilot");
     },
-    metadata: AGENT_COPILOT_CONTEXT_SERVER,
     tools_arguments_requiring_approval: undefined,
     tools_retry_policies: undefined,
     timeoutMs: undefined,
+    metadata: AGENT_COPILOT_CONTEXT_SERVER,
   },
   agent_copilot_agent_state: {
     id: 1023,
@@ -1497,10 +1477,10 @@ export const INTERNAL_MCP_SERVERS = {
     isRestricted: ({ featureFlags }) => {
       return !featureFlags.includes("agent_builder_copilot");
     },
-    metadata: AGENT_COPILOT_AGENT_STATE_SERVER,
     tools_arguments_requiring_approval: undefined,
     tools_retry_policies: undefined,
     timeoutMs: undefined,
+    metadata: AGENT_COPILOT_AGENT_STATE_SERVER,
   },
   // Using satisfies here instead of: type to avoid TypeScript widening the type and breaking the type inference for AutoInternalMCPServerNameType.
 } satisfies {
